+++ CHANGELOG.md
@@ -1,19 +1,23 @@
 # Changelog
 
 ## Next
+
 ###features
 - fixes the double header issue (#17)
 
 ## 0.7.1
+
 ###breaking change
 - opentelemetry updated to version 0.13. **Careful!!!** The opentelemetry version in your project should match the one in this library
 - several dependencies updated
 - docker file based on Rust 1.51
 
 ## 0.7.0
+
 ###breaking change
 - opentelemetry updated to version 0.12. **Careful!!!** The opentelemetry version in your project should match the one in this library
 - several dependencies updated    
+
 ###bugfix
 - tokio and tokio-test are now dev-dependencies