+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 # Changelog
 
 ### Next
+
+### 0.7.2
 #### features
 - fixes the double header issue (#17)
 - adds support for the ```gzip``` feature, which decompress the response content based on the content-type header
 - proper handling of graphql errors with a new `ParsedGraphqlResponse` type
+- adds ability to specify a request timeout
 
 ### 0.7.1
 #### breaking change